Update ListOfCourses.java
Added a method, contains, whose role is to loop through each course the user has taken and check if they have selected all each course's prerequisites. 

(For example, it is not possible to take CS 310 without taking CS 112, CS 211, and MATH 113. If the user forgets at least one of these, they will be unable to proceed)